No bug - Allow Android background tests to compile with -Xlint:all. r=me

+++ test/src/org/mozilla/gecko/background/healthreport/TestHealthReportDatabaseStorage.java
@@ -386,7 +386,7 @@ public class TestHealthReportDatabaseStorage extends FakeProfileTestCase {
     db.insertOrThrow("events_integer", null, v);
     db.insertOrThrow("events_integer", null, v);
     assertEquals(16, getTotalEventCount(storage));
-    final int nonExistentEnvID = (int) DBHelpers.getNonExistentID(db, "environments");
+    final int nonExistentEnvID = DBHelpers.getNonExistentID(db, "environments");
     assertEquals(1, storage.deleteDataBefore(storage.getGivenDaysAgoMillis(8), nonExistentEnvID));
     assertEquals(14, getTotalEventCount(storage));
 
@@ -436,7 +436,7 @@ public class TestHealthReportDatabaseStorage extends FakeProfileTestCase {
     db.insertOrThrow("events_integer", null, v);
     db.insertOrThrow("events_integer", null, v);
     assertEquals(16, getTotalEventCount(storage));
-    final int nonExistentEnvID = (int) DBHelpers.getNonExistentID(db, "environments");
+    final int nonExistentEnvID = DBHelpers.getNonExistentID(db, "environments");
     assertEquals(1, storage.deleteEnvAndEventsBefore(storage.getGivenDaysAgoMillis(8), nonExistentEnvID));
     assertEquals(14, getTotalEventCount(storage));
